% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_emails.R
\name{clean_emails}
\alias{clean_emails}
\title{Clean and augment a list of emails}
\usage{
clean_emails(emails)
}
\arguments{
\item{emails}{A list of DCOM email objects.}
}
\description{
Returns a tibble in which the first column is the given list of emails (as
objects), and the other columns are common attributes of emails.
}
