% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_file.R
\name{data_to_file}
\alias{data_to_file}
\title{Convert a data frame or tibble to a file}
\usage{
data_to_file(data, file_format = "csv", ...)
}
\arguments{
\item{data}{A data frame or tibble to be converted into a file.}

\item{file_format}{A file_format which will determine how the data is saved
and the extension of the resulting file.
Defaults to "csv" (comma-separated file)}
}
\description{
Convert a data frame or tibble to a file in a temporary location, and return
the file path. The file name will be based on the variable name of the
object, and will be given an extension that corresponds to the method by
which the file was saved. The `file_format` can be specified as "csv" 
(comma-separated file) or "tsv"/"txt" (tab-delimited). Support for Excel 
output is planned.
}
