% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_html.R
\name{data_to_html}
\alias{data_to_html}
\title{Convert a data frame or tibble to a HTML table}
\usage{
data_to_html(data, border = FALSE)
}
\arguments{
\item{data}{A data frame or tibble to convert.}

\item{border}{A Boolean value which, if set to TRUE, puts a border around the
HTML output. 
Defaults to FALSE.}
}
\description{
Uses the xtable package to convert a data frame or tibble to HTML, so that it
can be inserted into the body of an email.
}
