% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_image_cid.R
\name{embed_image_cid}
\alias{embed_image_cid}
\title{HTML tag to embed an image email attachment with a CID source}
\usage{
embed_image_cid(file_path, max_height = 800, max_width = 800)
}
\arguments{
\item{file_path}{Path of an existing image file.}

\item{max_height}{The maximum height of the image, in pixels. The image will 
be scaled down to the largest size that meets this contraint whilst 
preserving the aspect ratio of the image.}

\item{max_width}{The maximum width of the image, in pixels. The image will 
be scaled down to the largest size that meets this contraint whilst 
preserving the aspect ratio of the image.}
}
\description{
Generate a HTML tag that uses a CID source to display an image. This is used
in emails for images, in which an attached image can be embedded into the
email body. If the image is a png, jpg or bmp, the file will be scaled down
to the provided `max_height` and `max_width`, preserving the aspect ratio.
If the file is a gif, tiff or webp, the file will be scaled down without
preserving the aspect ratio.
}
