% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_to_file.R
\name{ggplot_to_file}
\alias{ggplot_to_file}
\title{Convert a ggplot to a file}
\usage{
ggplot_to_file(gg, device = "png", ...)
}
\arguments{
\item{gg}{A ggplot to be converted into a file.}

\item{device}{An argument passed onto the ggsave function that determines
how the plot is saved.
Defaults to "png".}
}
\description{
Convert a ggplot to a file in a temporary location, and return the file path. 
The file name will be based on the variable name of the object, and will be 
given an extension that corresponds to the method by which the file was 
saved. The `device` (eg. "png" or "jpeg") is passed to the `ggsave` function.
}
