% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_format.R
\name{image_format}
\alias{image_format}
\title{Return the image format of an arbitrary file}
\usage{
image_format(file_path)
}
\arguments{
\item{file_path}{File path of a file that may or may not be an image.}
}
\description{
Determines if a given file is a png, jpg, bmp, gif, tiff or webp image,
based on the file signature. If the file is none of these, NULL is returned.
If the file does not exist, an error is thrown.
}
