% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_attachments.R
\name{save_attachments}
\alias{save_attachments}
\title{Save an email's attachments}
\usage{
save_attachments(email, target_dir = tempdir())
}
\arguments{
\item{email}{A case-sensitive search query.}

\item{target_dir}{The folder in which to save the attachments. If no folder
is specified, a temporary directory will be used.}
}
\description{
Saves the emails from a given mailItem DCOM object (from RDCOMClient) to a
given folder. If no folder is specified, a temporary directory will be used.
This function will return a vector of all file paths to which an attachment
has been saved. The attachments will keep their file names as they are 
displayed in the email. Embedded images, such as those that are present in
email signatures, will also be saved.
}
