% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_emails.R
\name{search_emails}
\alias{search_emails}
\title{Search for emails}
\usage{
search_emails(search_term, folder = "Inbox", scope = "subject",
  partial_match = TRUE, search_subfolders = TRUE, search_time = 10)
}
\arguments{
\item{search_term}{A case-sensitive search query.}

\item{folder}{A folder in which to search.
Defaults to "Inbox".}

\item{scope}{The scope of the search: one of "subject", "body", 
"attachment_names", "from_name", "from_email", "cc_name", or "to_name".}

\item{partial_match}{If set to TRUE, will allow any number of characters to
appear either side of the search term.
Defaults to TRUE.}

\item{search_subfolders}{If set to TRUE, will search subfolders below the
specified `folder`.
Defaults to TRUE.}

\item{search_time}{The amount of time, in seconds, to wait for the search to
complete.}
}
\description{
Searches for all emails satisfying a search term. Searches can be performed
in one of several email attributes, such as subject or sender name. Searches
are always case sensitive. The DCOM search method used is asynchronous, in 
that R will not wait for the search to complete before continuing. We pause
R for a given amount of time (default is 10 seconds) to allow the search to
complete. See Details for more information.
}
\details{
There is a package for handling the AdvancedSearchComplete, RDCOMEvents, but
it is not easily available. The proper way to implement this search would be
to wait for the event to return that the search is complete, and then 
continue.
}
