% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_file.R
\name{data_to_file}
\alias{data_to_file}
\title{Convert a data frame or tibble to a file}
\usage{
data_to_file(data, file_format = "csv", file_name = NULL,
  col_names = TRUE)
}
\arguments{
\item{data}{A data frame or tibble to be converted into a file.}

\item{file_format}{A file_format which will determine how the data is saved
and the extension of the resulting file. Currently supports "csv" (comma-
separated), "tsv" (tab-separated), or "xlsx"/"excel". 
Defaults, "csv" (comma-separated)}

\item{file_name}{Sets the name of the saved file, without the extension. If
this isn't provided, the name of the `data` variable will be used. "." is a
forbidden `file_name` for this function, and is usually provided by a pipe 
(%>%); in this case, the file name will be changed to "data_", followed by a 
string of random numbers.}

\item{col_names}{Determines if column names (headers) are to be included.
Defaults to TRUE.}
}
\description{
Convert a data frame or tibble to a file in a temporary location, and return
the file path. The file name will be based on the variable name of the
object, and will be given an extension that corresponds to the method by
which the file was saved. The `file_format` can be specified as "csv" 
(comma-separated file), "tsv"/"txt" (tab-delimited), or "xlsx".
}
