% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_email.R
\name{prepare_email}
\alias{prepare_email}
\title{Prepare an outgoing Outlook email}
\usage{
prepare_email(embeddings = NULL, body = "", to = "", cc = "",
  subject = "", attachments = NULL, css = "", send = FALSE,
  data_file_format = "csv", image_file_format = "png")
}
\arguments{
\item{embeddings}{An optional list of all objects to be embedded into the 
email body or, in the case of a file path, the file will be attached instead, 
to allow files to be piped into the function. ggplots will be converted to 
images and data frames/tibbles will be converted to html.}

\item{to}{A string containing the email addresses, separated by
semicolons, which are to receive the email.
Defaults to an empty string.}

\item{cc}{A string containing the email addresses, separated by
semicolons, which are to be CC'd in on the email.
Defaults to an empty string.}

\item{subject}{A string containing the subject line of the email.
Defaults to an empty string.}

\item{attachments}{A ggplot, data frame, tibble, or file path which will be
attached to the email (converting to a file if necessary), or a vector 
containing any combination of these things. ggplots will be converted to 
images and data frames/tibbles will be converted to delimited files. 
If providing a list of ggplot or data frames/tibbles, it it best to provide
a named list. Otherwise, the attachments will be named attachment_1, 
attachment_2, etc.}

\item{css}{An optional string of CSS that will modify the HTML body of the 
email. Note that only some CSS affects Outlook emails.}

\item{send}{A Boolean value which, if set to TRUE, will send
the email immediately. If set to FALSE, will display the email on the user's
screen, but not send.
Defaults to FALSE.}
}
\description{
Prepare an email in Microsoft Outlook, either sending the email immediately
or causing the window to appear on the user's screen. This function prepares
emails with DCOM, which is only available in Windows, and is exposed to R
through the RDCOMClient package. Attachments are supported, as well as 
"embeddings"---a vector of data frames, tibbles or ggplot objects that are
embedded into the body of the outgoing email. It is recommended that 
Microsoft Outlook be open in the background before using this function.
}
\details{
No argument is mandatory; running prepare_email() will cause a blank 
composing window to appear. CSS can be entered with the css argument, but
note that most CSS is disabled by Microsoft Outlook.
}
