% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_file.R
\name{to_file}
\alias{to_file}
\title{Convert an object to a file}
\usage{
to_file(obj, ...)
}
\description{
Convert a ggplot, data frame or tibble to a file in a temporary location, and 
return the file path. The file name will be based on the variable name of the
object, and will be given an extension that corresponds to the method by
which the file was saved. If a file path corresponding to an existing file 
is given, then that file path is returned.
}
